Fix images in REPL for remote back ends
The basic problem was needing to use `file-local-copy` on the remote
image file.

Getting things set up to where I could test this, was not as simple as
I expected. I updated some configuration docs with an explanation --
maybe not entirely satisfactory? I need to do more research but I
think the status quo with Racket is there are some packages that
depend on `gui-lib` -- whether obviously or surprisingly -- that you
would want to use even if you do not want to run graphical desktop
apps on a remote headless server. To avoid the whole "cannot
instantiate racket/gui/base a second time in the same process"
rigmarole, we do need to load racket/gui/base eagerly. And in some
cases people would need to "xvfb-run racket".

I think this commit has the bases covered in terms of raw options for
people to use, but... I wish I could make the configuration experience
simpler/better, modulo what Racket allows.

Also I think a future commit might do well to flip the polarity and
change the `never-gui` property to `gui`.