Change back end configuration
Move back end housekeeping to its own new file, racket-back-end.el.

Change racket-back-end from a buffer-local variable to a function that
is called whenever we need a property list value.

The old way, each buffer would need to setq-local the var, which was a
nuisance.

Worse, it left the configuration "baked into" each buffer -- it
wouldn't be refreshed until the buffer re-visited a file or was
recreated. This was super confusing when experimenting with
configuration changes, but buffers would still be using the old
values tucked away into the old racket-back-end buffer-local var.

Now, with this commit, every use of a property list value is
effectively an on-demand function from `default-directory' to the
property list, so it will always be "fresh" and reflect the latest
configuration.

Also change the customization variable racket-back-end-function (a
function) to racket-back-end-functions (a list of functions, each of
which is called until one returns a non-nil value, the property list
to use. Although this might be overkill -- I haven't found it
necessary so far dogfooding some fairly non-trivial configurations --
it seems like a reasonable capability to provide.