Replace some cl-labels with cl-flet
Although harmless to use the former it is overkill in situations where
we don't need recursive bindings.

Both produce _lexical_ bindings, which is all we need in the
situations changed in this commit.

Note that a previous commit replaced a cl-flet with cl-letf +
symbol-function, because a _dynamic_ binding was needed for that
situation (causing a specific call to copy-directory to use copy-file
instead of make-symbolic-link).

Assuming I understand correctly:

- Where in Racket you'd locally define a "helper function" within some
lexical scope of another function definition, using let + lambda or an
internal define: In Emacs Lisp cl-flet is the equivalent (and
cl-labels if you need recursion).

- Where in Racket you'd use parameterize to change a function value
for some dynamic extent: In Emacs Lisp cl-letf + symbol-function is
the equivalent. (Sort of. This assumes in Racket you defined a
parameter whose value is a function, for example current-print. The
ability to change a function for some dynamic extent is something that
needs to be designed in ahead of time. Generally you can't just
"monkey patch" any arbitrary global binding in Racket, especially if
the binding is provided by another module, like you can in Emacs
Lisp.)