Use text lines data structure supplied by mflatt
This supports "paragraph" (lines) methods that are faster (e.g. about
0.6 the time) than those supplied by racket:text%.

Alas, our indent performance is still about 15X that of racket:text%.
Doing some preliminary benchmarks, this seems mostly due to
interval-map-ref/bounds being about 10X slower, and therefore the
other textoid methods being slower.

As a result, it may be that we need to replace interval-map.
Possibilities:

- Use token-tree% after all -- assuming it can support terminating a
re-lex early and efficiently shifting remaining positions, like we
were able to do with interval-map?

- Some bespoke data structure like text-lines. Either modify it also
to carry tokens alongside the strings, or, make some separate similar
data structure.

- Some other bespoke data structure. Don't necessarily rule out
something "dumb" like a gap buffer on a gvector; but probably it will
turn out to be some flavor of tree.

Having said all that I'm not sure that optimization is the best
next-action, as opposed to testing/improving things on the Emacs side
of the fence.